fix: address Eric's review — remove imapflow, extract shared util, CWD consistency
- Remove unrelated imapflow dependency from package.json/bun.lock
- Extract parseProcStatStarttime to shared util/proc.ts (was duplicated
  in runtime.ts and restart.ts)
- Store devCwd in PID file so restart.ts uses the same cwd hash as the
  watchdog, avoiding mismatches when SYNERGY_CWD is unset or differs
- vite.js SDK alias now checks all 3 dist entry files (index.js,
  client.js, server.js) to avoid partial build false positives
- requireSourceCheckout now also verifies packages/app and
  packages/synergy directories exist
- Updated test to reference util/proc.ts for parseProcStatStarttime
