+++ package.json
@@ -68,7 +68,6 @@
     "@ericsanchezok/synergy-plugin": "workspace:*",
     "@ericsanchezok/synergy-script": "workspace:*",
     "@ericsanchezok/synergy-sdk": "workspace:*",
-    "imapflow": "1.3.2",
     "typescript": "catalog:"
   },
   "repository": {
