+++ packages/synergy/src/server/runtime.ts
@@ -6,6 +6,7 @@ import { Instance } from "../scope/instance"
 import { Scope } from "../scope"
 import { ProcessRegistry } from "../process/registry"
 import { Log } from "../util/log"
+import { parseProcStatStarttime } from "../util/proc"
 import { InstanceBootstrap, ChannelBootstrap, HolosBootstrap } from "../project/bootstrap"
 import * as ChannelTypes from "../channel/types"
 import { Provider } from "../provider/provider"
@@ -39,22 +40,6 @@ async function runWithRestartPolicyAlways(options: RuntimeOptions): Promise<neve
   // as the parent, regardless of SYNERGY_CWD or later chdir() calls.
   const parentCwd = process.cwd()
 
-  // Parse starttime (field 22) from /proc/<pid>/stat output.
-  // Field 2 (comm) is parenthesized and may contain spaces, so we
-  // split from the last ')' rather than naively splitting on spaces.
-  function parseProcStatStarttime(stat: string): number | undefined {
-    const lastParen = stat.lastIndexOf(")")
-    if (lastParen === -1) return undefined
-    const afterComm = stat.slice(lastParen + 2).trim()
-    const fields = afterComm.split(" ")
-    // After comm, fields are: state, ppid, pgrp, session, tty_nr, tpgid,
-    // flags, minflt, cminflt, majflt, cmajflt, utime, stime, cutime,
-    // cstime, priority, nice, num_threads, itrealvalue, starttime (field 22)
-    // That's field index 19 in the afterComm split (0-indexed)
-    const starttime = parseInt(fields[19], 10)
-    return isNaN(starttime) ? undefined : starttime
-  }
-
   let shuttingDown = false
   let devRestartRequested = false
   let crashCount = 0
@@ -79,7 +64,7 @@ async function runWithRestartPolicyAlways(options: RuntimeOptions): Promise<neve
       // Store PID + startup identity for verification.
       // On Linux, include the starttime jiffies from /proc/self/stat so
       // restart.ts can verify the process identity without knowing CLK_TCK.
-      let identity: Record<string, unknown> = { pid: process.pid, startTime: Date.now() }
+      let identity: Record<string, unknown> = { pid: process.pid, startTime: Date.now(), devCwd }
       if (process.platform === "linux") {
         try {
           const selfStat = await Bun.file("/proc/self/stat").text()
